// frontend/src/api/client/models/user.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BotDashboard API
 * API per la gestione del Bot Dashboard con telemetria, comandi e statistiche real-time
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface User {
    'id'?: number;
    'userName'?: string | null;
    'normalizedUserName'?: string | null;
    'email'?: string | null;
    'normalizedEmail'?: string | null;
    'emailConfirmed'?: boolean;
    'passwordHash'?: string | null;
    'securityStamp'?: string | null;
    'concurrencyStamp'?: string | null;
    'phoneNumber'?: string | null;
    'phoneNumberConfirmed'?: boolean;
    'twoFactorEnabled'?: boolean;
    'lockoutEnd'?: string | null;
    'lockoutEnabled'?: boolean;
    'accessFailedCount'?: number;
    /**
     * @deprecated
     */
    'description'?: string | null;
    'admin'?: boolean;
    'lastLogin'?: string | null;
}