// frontend/src/api/client/models/create-user-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BotDashboard API
 * API per la gestione del Bot Dashboard con telemetria, comandi e statistiche real-time
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface CreateUserRequest {
    'username': string | null;
    'email': string | null;
    'description'?: string | null;
    'isAdmin'?: boolean;
    'password': string | null;
}